+++ android/app/src/main/kotlin/cz/standakouba/rozhledny/MainActivity.kt
@@ -1,4 +1,4 @@
-package cz.rozhledny.rozhledny
+package cz.standakouba.rozhledny
 
 import io.flutter.embedding.android.FlutterActivity